requirements.txt
streamlit
yt-dlp
openai-whisper
setuptools